[CpmlSegment] Rewrote API almost from scratch
The old API was duplicating the path concept yet present on cairo
and the origin (kept in its own struct and detached from the cairo
path) did not allow interchange between cpml and cairo.

The new API is by far cleaner, using less functions, and it is a
real complement to the cairo path, not a substitute. The CpmlSegment
is now an internal structure that helps to modify the original
cairo_path_t passed in cpml_segment_init(), the cpml entry point.