[CpmlPair] Implemented cpml_pair_at_... family
adg_pair_at_line() uses the "t" parameter as percentage between the
starting and the ending point, where 0 is the start and 1 is the end
(being linear, 0.5 means the mid point).

At the reverse side, adg_pair_at_curve() uses "t" as the "time",
that is the value of "t" in the canonical Bézier function. Again,
0 is the start point and 1 is the end point but 0.5 is not
necessarely the mid point.