[CpmlSegment] Initial implementation of cpml_segment_offset()
The offset API is the fundamental base for a lot of CADG operations.
This first (rough) implementation works for lines but not for curves:
the control points need to be arranged. Now they are left untouched
as I did not found a decent way to approximate the cubic Bézier curve.