[CPML] Added stub for AdgPrimitive and AdgCurve
The plan is to use the AdgPrimitive as an abstraction layer that
dispatches the operations common to all the primitives to the
relevant one. This will simplify a lot the job of AdgSegment (as
almost all the operations should be propagated to the underlying
primitives) and will help future development, though.