[CpmlClose] Completed API for close path lines
Added the missing APIs: cpml_close_{pair_at,vector_at,offset}().
The implementations are practically the same of those ones used
for the line primitive, but using cpml_primitive_get_point(-1)
as end point instead of a specific point number (as the only way
to cycle the segment is requesting the end point of a close
primitive: check the cpml_primitive_get_point() documentation).