[CpmlPrimitive] Improved cpml_primitive_get_point()
Added the negative npoint feature: if npoint is negative, it is
considered as an index from the end point, meaning -1 is the end
point, -2 the point before the end point and so on. This is an
addition to the previous meaning, where a negative npoint was only
specifying the end point.