[CpmlAction] Renamed cpml_join_primitives() to cpml_join()
Renamed the join action to be consistent with cpml_intersection().
Moved on its own source file, so now cpml-action.c contains only
comments to properly generate the action section documentation.
This also was needed to be consistent and because the joining
of primitives involving curves (actually still not implemented)
will be quite challenging.