[AdgPath] Slightly modified the behaviors of the manipulation API
The CPML library is quite green and requires a good path quality,
so the AdgPath manipulation APIs have been modified from the
original cairo function behavior in a more restrictive way.

For example, following the rule of the less surprise, some
cairo functions guess the current point when it is not defined,
while the AdgPath methods trigger a warning without other effect.
Furthermore, after a cairo_path_close_path() call a MOVE_TO
primitive to the starting point of the segment is automatically
added by cairo while in the ADG, after an adg_path_close(), the
current point is simply unset.