[AdgPath] Refactored AdgPath to support arc primitives
Implemented Bézier curve approximation of arcs during
adg_path_get_cairo_path() and reworked the clear operation of the
internal cairo path struct to avoid code duplication.

The curve approximation is using an harcoded value as max angle
(PI/2): it is planned to implement a property providing this value.