[build] Included --enable-gtk2 in configure
This option allows or not the inclusion of GTK+ based widgets inside the
ADG library (libadg). If no value is explicitely specified, it defaults
on an automatic check, that is it looks for any usable gtk+-2.0 package
in the system. Furthermore, following the rule of the least surprise, if
--enable-gtk2 is specified but no valid gtk+-2.0 package is found,
configure will hang.