[AdgEntity] Improved "invalidate" signal
Using "invalidate" about the same way the "render" signal is managed,
but revering the meaning. This means the default "invalidate" handler
is not overridable but instead call the render() virtual method.

If render() is not implemented, a valid empty invalidate callback is
assumed (useful for entities without cache). At the end, if the callback
run succesfully (that is, returned TRUE) the RENDERED flag is reset.