[AdgEntity] Changed "parent" type to AdgEntity
Previously the "parent" property was AdgContainer. Changing to AdgEntity
provides a cheap way to define dependencies between AdgEntity: this is
required, for instance, to render quotes using AdgToyText internal
entities without forcing AdgDim to be a container.

This change involved a lot of recoding: actually the AdgEntity owns a
reference to the parent entity. In the usual case, that is when parent
is an AdgContainer, it owns a reference to the child too, so you'll end
up with a circular dependency between parent and child.

The code has been updated to meet the new AdgContainer code.