[AdgDim] Using internal AdgToyText for labels
This feature required a rewrote almost from scratch of this object and a
bunch of improvements previously implemented.

AdgDim now embeds 4 internal AdgToyText:
1. basic value: the ideal value for this quote
2. min value: style dependent, could be low tolerance or low value
3. max value: same as above, but in the other side
4. note: a custom text to be appended to the quote

The API has been renamed to make clear the difference between "quote"
and "basic value": the former is the collection of all the AdgToyText
while the latter is only one of them.