[AdgTrail] First implementation
Splitted the old AdgPath in AdgTrail (parent) and AdgPath (child). The
former is a generic CpmlPath model that demands all the implementation
to the caller.

This is needed for path models that are subject to different
transformations on a point basis, such as what needed to draw extension
and base lines for AdgLDim.