[AdgDimStyle] Implemented marker customization
Added a powerful way to customize markers for dim-styles by using
template entities. The style provides two different markers: the fact
that "marker1" is the starting and "marker2" is the ending one is a
mere convention.

When adg_dim_style_use_marker[12]() API is called, the property
values of the passed-in marker is done and hold internally. Whenever a
new marker is requested (throught adg_dim_style_marker[12]_new()),
a new object is created and initialized with these properties.