[AdgDimStyle] Renamed tolerances to limits
Updated APIs to reflect the change: "limits" is a more generic term than
"tolerances". In the future, alternative dimensioning styles could be
supported and some of them use min/max values instead of basic value
with tolerances.

Also, renamed down/up to min/max to be consistent with AdgDim and
removed the "note" management. Eventual note text will be implemented
with custom quote value and using the number tag where the quote need
to be substituted.