[AdgEntity] Added the arrange infrastructure
The render() has been splitted in two different operations: arrange()
and render(). While the latter performs the usual rendering to a cairo
context, the former provides a way to arrange the layout of the entity
components without accessing the cairo context. This will be used for
the extents computation.

Implemented the "arrange" signal, provided its accessor and added the
adg_entity_get_extents() API (and the setter, to be used internally)
that update the entity emitting "arrange" just before computing the
extents.