[AdgToyText] Forcibly unset the font in global-changed
When changing the global map, the font could be recomputed if this
change involves something deeper than a translation. The font caching
(that is avoiding the font recomputation when not needed) is done in the
AdgFontStyle class: there is no sane reason to double the caching
algorithm also inside the "global-changed" signal.