[AdgEntity] API cleaned up
arrange() and render() virtual methods must have a common API, so
adg_entity_apply_local_matrix() has been dropped. Changed the accessors
to the global and local matrices to return a const pointer to the
internal matrix, hence avoiding a call to adg_matrix_copy() and allowing
to remove the setter (adg_entity_set_local_matrix() and
adg_entity_set_global_matrix()). In the worst case, if an implementation
needs to change on-fly a matrix can cast the returned const pointer from
adg_entity_{global,local}_matrix() and modify it.