[AdgLDim] Use only one "pos" property
Updated the code accordling to the previous commit. The double pos was
born because ADG was using "pos1" and "pos2" as the points above which
the dimension was growing in local space and below which it was growing
in global space (usually a point on the boundary box of the model).
Now the position specifies (correctly) the same things but only for the
quote: the double points are calculated internally and not exposed.