[AdgADim] Corrected arc behavior
Code clean up: update_geometry() and update_shift() are now merged
togheter in update_geometry() as they share a lot of code. Split the
basic geometry stuff in get_info() (as once upon the time done for
CpmlArc) to not have a never-ending function.

The reverse arc problem was a distance of 0 (a not so remote case, as it
comes out whenever the pos reference is the center of the dimension):
added a minimum setting to avoid the 0 case... a quite naive approach
but working great.