[AdgFillStyle] Initial implementation
This class will be the base class for all the fill styles. Above it
will be built at least the solid fill (a solid color that maybe in the
future could overtake the AdgColorStyle object), ruled fill (heavily
used in technical drawings to show sections) and custom fill (that will
depend on a user provided callback).