[CPML] Implemented the extents infrastructure
Added the extents function to every primitive type. Usually, calculating
the extents of a primitive is quite trivial (just add the pairs to an
initially undefined extents throught cpml_extents_pair_add()), but the
most complex primitive to be managed is the arc.

Regardling the Bézier curve, the control points are added as is: the
control poligon always contains all the Bézier curve but the result
could be more than what expected.