[ADG] Removed adg_fill_style_add_extents()
Directly using adg_fill_style_set_extents() from AdgHatch to set the
AdgRuledFill max extents. The previous approach suffered from the lack
of reset, thus letting grow the extents also with a simple translation
of the matrix. Now a single pattern covers the bigger area and is
translated for every area to hatch.