[AdgEntity] Changed "local-mode" to "normalized"
The local matrix must be applied always before the global one, so there
is no need to hold the full-fledged "local-mode" property: changed
it to the gboolean "normalized".

The adg_entity_get_ctm() API has been renamed to adg_entity_cpm() that
returns a const AdgMatrix * instead of storing it somewhere. This needed
to hold a new AdgMatrix in the AdgEntityPrivate struct.

The ctm update is now performed in the arrange default signal handler
prior to call the custom signal handlers.