[AdgPath] Added API to access last and over primitives
Implemented adg_path_last_primitive() and adg_path_over_primitive() to
get respectively the last and the previous the last primitives appended
to an AdgPath. Modified the code accordling and let the default append
method (adg_path_append()) do his job also when adding the result of an
operation (such as a line for chamfers and an arc for fillets).

The adg_path_over_primitive() is especially useful to access the
primitive added by the forementioned operations. The _last version has
been added only for completness.