[AdgModel] New API to manipulate the named pairs
The new adg_model_foreach_named_pair() API allows to execute a
callback programmatically (reflecting what GLib does for has tables).
Returning a list of pairs is easily implementable by using this
function and passing an empty GSList as used_data, so an API to do this
is not planned.