[AdgPath] Added adg_path_reflect() API
New function to reflect the first segment of an AdgPath around an
arbitrary axis, joining the resulting segment to the original one. It is
quite useful when drawing turned parts, always symmetric respect the
turning axis.

Improved error management by adding the G_STRLOC field and localizing
the message with the _() macro.