[AdgEntity] Using AdgLocalMode instead of AdgTransformMode
Changed the "local-mode" type to the newly AdgLocalMode one. This avoids
the need to post process the ctm because now it is ensured the local
matrix is always known and applied before the global one. Furthermore,
using AdgLocalMode allows more advanced customization on the local
matrix computation, such as what done by the ADG_LOCAL_FROM_PARENT
method.