[AdgTable] Added frame support
Now every table cell has its own frame flag: this means the frame can be
enable or disabled on a cell basis. Added a similar feature (but with
its proper GObject property) to the table.

Actually only the whole frame could be switched on or off: it is planned
to implement single line switching only when (and if) will be required.