[AdgEntity] Matrix recomputation only in arrange()
Updating the local and global matrices is done by the "local-changed"
and "global-changed" default signal handlers. These signals are now
emitted by the real_arrange() default signal handler and only if their
own is_defined flag is TRUE (set whenever the relative map is changed).

There is no need to recursively set the is_defined flag on the children
because the "local-changed" and "global-changed" signals are the ones
forwarded to the descendants.

Closes issue #11: http://dev.entidi.com/p/adg/issues/11/