[CpmlPair] Return nothing from pair setters
Do not return CpmlPair/CpmlVector from cpml_pair_copy(),
cpml_pair_from_cairo() and cpml_vector_from_angle() to make clear these
functions do not allocate a new pair but set a pre-existing struct.

The pointer was returned in order to minimize typing with a basic
"fluent" approach but it came out it was misleading: issue #23 provides
a good example.