[AdgTrail] Avoid recursion caused by bad callbacks
If the user provided callback tries to access the CPML path by
explicitely or implicitely generated call to adg_trail_cpml_path(),
there will be a mutual recursion between the callback and the
cpml_path() virtual method.

Added a guard flag to adg_trail_cpml_path() to avoid this situation.