[CPML] Wrapped CAIRO_PATH_... constants
Using specific CPML_... constant instead of cairo ones. This was not
strictly necessary but it is needed in order to have at least one symbol
to document for every primitive (to make gtk-doc happy) and to be
consistent with the arc primitive, the only one currently undefined in
cairo.

Updated the whole project to reflect this change.