[build] Make LT_INIT macro optional
LT_INIT has been defined since ages (2004) but it is a macro available
only on post 1.6 versions of libtool and some distribution (most notably
slackware) still use libtool 1.5.

This commit still uses LT_INIT but it will fallback to the good old
AC_PROG_LIBTOOL if it is not defined.  Patch tested on a vanilla
slackware 13.0.