[CpmlSegment] Removed dependencies on n_points
The primitive size does not depend on the number of points a primitive
may have but only on the "length" field. This because primitives could
embed additional data in the holes between the end of their points and
the effective data length.

The CpmlSegment APIs have been cleaned up and redesigned to deal with
this issue, that is without assuming anything about the data length.
Furthermore, their implementations are now more robust do to a strictly
check of CpmlSegment compliance: a segment must be contain exactly one
CPML_MOVE element, and only at its beginning. Failing to fulfill these
rules will result in a failing cpml_segment_from_cairo().

Closes issue #21: http://dev.entidi.com/p/adg/issues/21/