[AdgMarker] Avoid CpmlSegment caching
Do not try to cache results while a new segment is requested: although
trail and n_segment could be the same, the internal CpmlSegment data
could change.

When AdgLDim arranges the layout and changes the path data (to force
outside arrows for example), it reuses the same CpmlPath. This means
trail and n_segment are still the same, but the internal CpmlPath data
could be totally different, hence causing segmentation faults in some
situations.

Closes issue #33: http://dev.entidi.com/p/adg/issues/33/