[tests] New adg_test_invalid_pointer() function
Common code to get an invalid pointer, that is a pointer to some random
memory. Using it in place of handcrafted "junk" array.

It is not advisable to use naive junk pointers such as (void *) 1: this
will likely point to read-only memory, hence breaking the tests because
of this instead of checking the reaction with wrong pointers.