[tests] New adg_test_add_func() common function
Alternative to g_test_add_func() (and compatible with it) that hides the
stderr output if the test function succeed but shows it when the test
fails.

This is done by always performing the test inside a forked environment
without stderr output: if the function fails, the test is rerun without
trapping stderr.