[build] Refactored to change installed behavior
Changed the installation to a (arguable) cleaner approach, using the
"cairo" way instead of the "gtk" way.

The main headers, that is the ones to be included by the projects, are
now in the root directory included by the pkg-config files, so the
project/project redundancy is avoided. In other words, an application
should now directly include <cpml.h>, <adg.h> or <adg-gtk.h>.

To avoid potential clash problems with other adg/cpml packages, the
headers are installed under adg-1 and cpml-1 and the libraries have been
renamed to libcpml-1 and libadg-1. The new pkg-config files are adg-1.pc
and cpml-1.pc.