[build] Do not hardcode the glade catalog path
Although the official glade documentation suggests it:

http://library.gnome.org/devel/gladeui/stable/catalogintro.html

hardcoding catalogdir is a bad idea in the first place and it breaks
VPATH builds. I have not found the ideal solution so I choosed an ugly
one: substituting ${datarootdir} of the glade catalog path with the one
set by the ADG project.

The implementation exactly tries to do this, although some corner cases
are not properly handled. For instance, if the glade catalog path does
not start with `pkg-config --variable=datarootdir gladeui-1.0`, it will
be hardcoded verbatim with the above problems.