[demo] Initial data manipulation
Started the real job on issue #3: the model manipulation by changing the
data. Actually, the shape is properly modified but the associated
entities (such as AdgEdge and AdgLDim) do not update accordingly.

The edit window has been modified to properly set the initial values
(instead of having an hardcoded list). "Apply" and "Reset" buttons are
sensitives only when appropriate.