[AdgPoint] Improved API and updated accordingly
Removed adg_point_set() because adg_point_copy() yet provides a similar
behavior. Implemented new adg_point_unset() API to make it clear how
unset an AdgPoint, avoiding to rely on a NULL value passed in as
parameter to other _set() methods.

Provided a convenient method in AdgEntity to set a point that must
update also the dependencies on the model (when that point is a named
pair) and used throughout the dimensioning entities.