[adg-demo] Hold last wndEdit position
The old one-time callback was setting the USER_POS hint after the window
was positioned the first time. This worked great for "delete-event" and
default "close" signal (that is, closing the dialog with the window
manager or pressing the [ESC] key), but I did not find a viable way to
do the same with inside a "response" answer (that is, after a click on
the "Close" button).

The new approach adds a callback to the "response" signal that stores
the window position, hides it and moves the window to the old position
again. This works flawlessy with openbox-3.4.11.1, and I really hope to
not touch this stuff again because the window manager interactions seems
to be a big mess...