[AdgDim] Properly use the "value" property
AdgDim:value is now used as template to generate the set value of the
quote. Inside this template string, the place holder specified by the
AdgDimStyle:number-tag (usually "<>") is substituted with the default
value for the quote.

This will allow to prepend, suffix or even set a custom string in place
of the default value.