[build] Cross-building on MinGW32 working
Updated and corrected bugs in the build system: now ADG cross compiles
out of the box in a MinGW32 system on a 64 bit platform and the adg-demo
runs properly on wine (although it seems to have serious problems while
rendering vertical text). On a native Windows XP platform these problems
desappear though.

Provided a wrapper (build/configure-mingw32) to configure the cross
building on an x86_64 Arch linux system.

Closes issue #34: http://dev.entidi.com/p/adg/issues/34/