[tests] Removed custom CFLAGS/LIBS and added $(EXEEXT)
Directly using CFLAGS and LIBS returned by pkg-config, avoiding to rely
on the custom flags set by configure: this will avoid duplication
library dependencies.

Added $(EXEEXT) to the binary test programs to make them available on
windows platforms. This is still not tested because a cross-platform
build forbids to execute the foreign binary. Although not so important,
an approach based on wine could be investigated.