[build] Improved runtime check of data files
Instead of hardcode srcdir and builddir, check the file using the caller
path and hacking around it.

If argv[0] contains LT_OBJDIR (usually ".libs"), assume the program has
been run uninstalled and check for data in the parent directory of the
called binary. Otherwise, use PKGDATADIR, defined to $(pkgdatadir) with
the preprocessor flags.