[build] Improved configure wrapper for MinGW
Renamed configure-mingw32 to configure-mingw: as stated on the home page
of the project, to "prevent the implication that MinGW will only works
on 32 bit systems".

Using a custom INSTALL_DIR and defining some prefixes to more useful
values: the installation will be performed in the top_builddir, under
the "_mingw" directory.