[build] Readded canonical host check
Checking the canonical host in prevision to implement different data
lookup approaches based on the host platform. I did not find a viable
way that works on every system: autoconf expects all prefixes to be
absolute paths while a windows installer (NSIS in this case) allows to
install the project tree wherever the user wants.