[AdgPair] Added adg_pair_copy()
New API to perform a value copy between two pairs. The same as
cpml_pair_copy() but with argument validation.

The rationale behind this is to keep the CPML library bloat free by
not checking anything and wrapping these functions inside libadg in a
more robust way only when this makes sense.