[AdgCanvas] Implemented imposed canvas size
Using the AdgCanvas:size property to impose that specific size when set.
If the size is not set (that is, if it is 0) the behavior is the old
one, that is the canvas size is determined by the contained entities.

Modified the adg-demo program to deal with the new feature.

Closes issue #40: http://dev.entidi.com/p/adg/issues/40/