[AdgTitleBlock] Prefixed internal symbols and improved "date" behavior
Prepended _ADG_ and _adg_ to static functions and local macros as
required by issue #32. Also, removed internal setters and relying on
GObject property built-in validation.

The AdgTitleBlock:date behavior has been slightly changed: actually you
need to set it to NULL to get the today date automatically rendered and
use an empty string to not have any date (the reverse as before).