[ADG] Avoid to forcibly set the CTM on the cr
The GTK+ printing API modifies the ctm on the cairo context accordingly
to the page orientation but the ADG canvas will forcibly set the current
transformation matrix, overwriting whatever set.

Modified all the render() methods to not set the ctm: they will properly
modify the ctm instead. This allows to set a custom ctm before the
rendering that affects everything, as requested by the printing API.