[demo] Backup/restore global map while printing
To emit the drawing for previewing and printing purpose, the global map
of the canvas must be set to the identity matrix. The old map is stored
and restored as soon as the drawing has been generated, so the artifacts
rendered on the screen due to wrong global map (highlighted in the
previous commit) are now vanished.

Closes issue #44: http://dev.entidi.com/p/adg/issues/44/