[AdgPoint] Do not raise warnings on adg_point_get_pair()
Avoid warning messages if the subject point is a named pair and the name
is not defined in the model. This is the first step to allow the
implementation of _optional_ entities that automatically disappear when
all the points needed are not fully defined.

Started issue #44: http://dev.entidi.com/p/adg/issues/44/