[AdgModel] Added "reset" signal
New signal to reset the model to the initial state. Actually, it
basically removes all the named pairs. Now it is possible to redefine
the model with optional parts: the named pairs are cleared, hence the
eventual entities bound to those pairs are not rendered.

The issue #44 is basically resolved with this commit but will be closed
as soon as the optional rendering of parts will be included in the demo.