[demo] Updated adg-demo to the latest changes
Implemented the rendering of the optional groove machining by leveraging
the new automatic skip of incomplete entities render. This means
dimensions are rendered or not depending on the presence of the groove
in the model data.

General improvements of the drawing by using Bézier curves instead of
straight lines to expose the rear hole. Furthermore, moved the hole data
outside the GtkExpander in the data dialog: the hole is always present
in the drawing and it is not optional.

Closes issue #44: http://dev.entidi.com/p/adg/issues/44/