[AdgText] Do not modify the local matrix for alignment purpose
The pango library uses the top/left corner as reference for text layout
objects. The cairo "toy" text API uses the bottom/left corner.

To be consistent, the AdgText entity rendering was displaced to match
the AdgToyText implementation by modifying the local matrix. The problem
was this displacement was cumulative and for every invalidate() signal
the AdgText entities were "floating" away.

The new implementation uses a manual translation in the render()
function, avoiding the abuse of the local matrix. Furthermore, corrected
a bug related to this displacement in the extents computation.