[AdgAlignment] Corrected displacement calculation
The use of the inverted global matrix is plain wrong. For example,
suppose to have a text label of 3 width x 1 height: if it is rotated by
45 degree, the new extents will be something around 2,5 x 2,5. Applying
the inverted global matrix will result in a 2,5 x 2,5 box, not the
original 3 x 1.

Reverting the logic to the old approach of staged arrangements,
implemented with cleaner code.