[AdgEdges] Enabled use of the critical-angle property
Although defined, AdgEdges:critical-angle was not used for computing the
edges. Changed its implementation by relying on an internal ghost
duplicate (as almost any other property) instead of elaborating the
value and storing the result, pulling in rounding errors. The threshold
is now calculated from this property when finding the edges.