[AdgEdges] Added the "axis-angle" property
Allow a custom axis for edge lines. This custom axis must pass through
the origin (0,0) but has a custom slope defined by the new
AdgEdges:axis-angle property.

The corners data are transformed before the computation to transform
this axis to y=0, so the old algorithm is still valid. After the
CpmlPath containing the lines is returned, the inverse of the original
transformation is applied to every coordinate to move the points to
their original position.