AdgRDim: implemented outside style
Enhanced AdgRDim by implementing an alternative style when the "outside"
property is enable.

The new dimension should render a leader line starting from a styleable
distance from the arc (in the opposite side of the quote) to the natural
end the current radius dimension, that is from (<---R to -->(---R where
the bracket is the arc.

Closes issue #9.