AdgEntity: added "destroy" signal
g_object_unref() is not enough to destroy a composite entity because
its children still hold references to itself. This is why the "destroy"
signal is needed for: it will be propagated to the children (or to its
internal entities) so at the end all the references will be released
and the entity will likely be destroyed.

Anyway the entity will be still alive after a destroy process if the
caller owns a reference to it.