build: introduced NTD_UNEXPAND() autoconf macro
Instead of getting the hardcoded paths from pkg-config variables, added
a macro that gets the (likely) unexpanded version. This is done by
redefining the commond used variable (prefix, bindir, libdir and so on)
to their unexpanded version ('$prefix', '$bindir', '$libdir')... kind of
hackish but the cleaner way I found of.

Used this macro to get glade catalogdir and GObject introspection dirs,
so now all the required directories are relative to '$prefix' and make
distcheck does not fail.