build: corrected MKINSTALLDIRS in po-properties/Makefile.in.in
The MKINSTALLDIRS var above suffers from serious expansion problems:
$ac_aux_dir, part of MKINSTALLDIRS when AC_CONFIG_AUX_DIR() is used,
is embedded in its expanded form. This means is useless because can be
either an absolute path or a path relative to $(top_srcdir).

To avoid tinkering with this path, the proper hardcoded version
"$(top_srcdir)/build/mkinstalldirs" will be used instead.