docs: merged and adapted gtk-doc 1.18 into ADG
Imported makefiles and stylesheet and adapted to ADG requirements.
Actually the generated documentation is parsed with XSLT for keeping the
website up to date, so it needs to be XHTML compatible. There is no way
to override that in gtk-doc, so gtk-doc.xsl (with relative dependencies)
has been imported and adapted.

The build system does not call gtkdoc-mkhtml anymore but uses an hand
crafted call to xsltproc to do the same conversion with the custom
stylesheet.