demo: relocated the default window icons
Instead of hard-coding PKGDATADIR in libadg, the function is now generic
and the path decision is demanded to the application (that has access to
its arg[0]). Modified and adapted the sources to properly handle the
icons and the .ui files.

Put some shared code in demo.[ch] to avoid duplication.