build: avoid gtester if cross-building for Windows
gtester requires a POSIX compliant system for signal handling, so it is
not built by default on Mingw: the relevant automake directives in GLib
are wrapped by "if OS_UNIX" and "endif". It is actually not possible to
execute "make check" by using wine.

Actually the check is disable so "make check" will be a noop.

Closes issue #43: http://dev.entidi.com/p/adg/issues/43/