adg: dropped ADG_TYPEDEF
This was a horrible hack to get gtk-doc working. Anyway it messed up the
GObject introspection side.

Dropped the hack and accessed a private struct in cairo.h (the lesser of
two evils). Rearranged code to still work with gtk-doc by using a helper
header file disabled only from the gtk-doc side.