adg: use cairo_matrix_t instead of AdgMatrix
An alias such as the old AdgMatrix creates problems with bindings.
Removed the alias and directly used cairo_matrix_t instead. Provided a
backward compatibility wrapper that implements a GBoxed around
cairo_matrix_t in the way cairo would implement it (cairo-gobject, for
some unknown reason, did not wrap it: see [1] for details).

[1] http://lists.cairographics.org/archives/cairo/2013-March/024104.html