adg: inherit AdgDress from GEnum
Instead of using a hand-crafted solution to provide AdgDress integers,
inherited a new type from GEnum and using it. This will make bindings a
lot easier because string to int mapping is yet provided.

Removed API to append new values to the AdgDress class (they will be
added when needed).