build: added _build/ to POTFILES.skip
Ok, this one is so shitty is not even funny. When in VPATH build under
a subdirectory of top_srcdir, make check is failing in po/ with:

    srcdir=../../po /usr/bin/intltool-update -m
    The following files contain [...bullshits cut...]

    [VPATH dir]/demo/adg-demo.ui
    [VPATH dir]/demo/cpml-demo.ui

    If some of these files are left out [...further bullshits...]

Now the above files are generated by configure: how am I supposed to
behave? I cannot add them to POTFILES.skip because the damned thingy
(AKA intltool) looks in all the subdirectory under top_srcdir.

To keep my sanity I added _build/ to POTFILES.skip and I will force
myself to use it (hoping to remember this) in VPATH builds.