adg: bare export API adg_canvas_export()
Added also the helper function adg_type_from_filename() that tries to
guess the surface type from the file name. Modified adg-demo to use the
new APIs accordingly.

WARNING: one argument of adg_canvas_export() is a cairo_surface_type_t.
Actually cairo does not export such type via gobject-introspection so
that API will not be available to language bindings. I made a pull
request to add support to that (and more) types:
https://github.com/GNOME/gobject-introspection/pull/1
The API will be available only when that PR will be merged upstream.

Closes issue #106: http://dev.entidi.com/p/adg/issues/106/