adg: correct margin handling
When the canvas is bound to a page setup, either because of
adg_canvas_set_page_setup() or adg_canvas_set_paper(), keep the paper
size constant, i.e. decrease or increase the *page* size to compensate
increases or decreases of the relevant margins.

Added more tests to check the margins behavior.

Closes issue #109: http://dev.entidi.com/p/adg/issues/109/