build: use autopoint 0.17 to provide gettext support
The problem is if I don't want to depend on external gettext, I need to
import inside the repository a bunch of stuff, including some M4 macros
that can possibly conflict with other build components.

If I depend on external gettext without providing a version, the build
fails because there is no po/Makefile.in.in template. The stupid
gettextize command requires user interaction, making it unsuitable for
automated scripts.

So the "solution" (pun intended) is to use autopoint (automatically
triggered by autoreconf) by feeding a specific version to
AM_GNU_GETTEXT_VERSION.

I opted to depend on 0.17 because it is old enough to be decently
supported on modern distros and still it is not a ten years old relic.