tests: avoid g_test_trap_fork() in tests
To avoid cluttering stdout with warnings, I was trapping every test
without logging enable, eventually redoing the last one if failed. This
required forking the process with g_test_trap_fork(), now deprecated
because it does not always work.

Adopted another approach: GLib 2.38 introduced a new API
(g_test_set_nonfatal_assertions()) that does not abort the program on
assertion failures. So now every test stores the warninings in an
accumulator and, when finished, if the test has failed that accumulator
regenerates is replayed to show the old warnings.

Closes issue #153: http://dev.entidi.com/p/adg/issues/153/